requirements.txt
Flask-HTTPAuth
Flask
Pyjwt
requests